Test move/deletions, one re-categorized.
The size of VLAs at prototype scope aren't evaluated and forming a composite
type of VLAs with unevaluated size expressions is UB, but VLAs in prototypes
(i.e. declarations that aren't definitions) are supposed to be treated as
having unspecified size (and therefore forming their composite types should
probably be well-defined).