refactor attributes (#296)
* refactor attributes

The old attributes were trying to convert everything into a K,
which led to a bunch of performance problems and also various
exceptions thrown when things couldn't be converted. We are now
just treating attributes as a map from string and class keys to
objects of the specified class. The implementation provides
convenience helpers for string attributes and class singletons.

This should significantly cut down on the amount of nasty scala code
needed to deal with attributes.

* fix injection bug

* fix expected output of tests

* fix javadoc for ExpandMacros

* try to adjust for nondeterministic errors

* Revert "try to adjust for nondeterministic errors"

This reverts commit 9588064ffaa54854d962ae67e03559868a44232c.

* remove more nondeterministic tests