delete unused profiles and create a second profile which reverses evaluation order (#317)
What this does is create a flag to kompile which reverses the order of rules prior to
compilation to OCAML. We also delete the 78k0r and x86-gcc-limited-libc profiles from
the installer since they aren't very good and aren't being used to my knowledge.
We then create a second profile which uses this flag. As a result, by using both profiles
you get much higher probability of detecting undefined behavior and test failures which arise
from extremely simple nondeterminism in your program.