Add setting to allow  __restrict and __restrict__ variants
This resolves the issue #345 that our previous
implementation by #define __restrict restrict
breaks with autoconf detection of restrict support, which
tests first whether __restrict works, and if it does will
emit #define restrict __restrict.

With this commit the lexer recognizes __restrict and __restrict__
as a new "RestrictReserved" token, which the translation semantics
either canonicalizes to the plain Restrict or signals as an error,
following the new Bool ::= cfg:reservedKeywordAllowed(String) setting.

All current profiles are based on gcc which allows these variant keywords,
so the error reporting code is not tested!