Make plain c-semantics (native) build work again
This fixes two problems.
The first is that the options that moved from kcc to
semantics-kompilation time should have been specified
in the makefiles under c-semantics rather than in the
arguments the c-semantics-plugin makefile passed to
those makefiles.

The other was fixing problems where the OCaml backend code
for linking the generated OCaml files was incorrectly
and inconsistently computing paths to intermediate files.
This was missed by Jenkins and in development because
the c-semantics-plugin/src/main/ocaml/Makefile specifies
this compilation and linking itself and passes the
--gen-ml-only flag to the backend.