Fix name mangling for constructors and destructors (#848)
The original code was causing an assertion failure inside clang,
since user is expected to use `mangleCXXCtor` and `mangleCXXDtor`
for mangling constructors and destructors. However,
when a release build of clang was used, the assertion was disabled
and clang-kast generated some mangled name for ctor and dtor.

In particular, for constructors, the mangled name for `Ctor_Complete`
was generated, and for destructors  `Dtor_Deleting`. The current code
uses `Ctor_Complete` and `Dtor_Complete`. However, I doubt it is correct.
We will probably need to mangle all variants of ctor and dtor.

Also see the section 5.1.4 of 'Itanium C++ ABI': http://refspecs.linuxbase.org/cxxabi-1.83.html