C++ Constant Expressions - Refactored (#863)
* Extract the definition of <thread-local> cell to separate file.

* In C++ translation semantics, rename 'curr-program-loc' to 'curr-tr-program-loc'.

This is needed because 'constexpr' support requires the 'thread-local'
cell in the translation semantics, and the 'thread-local' cell already contains
a cell named 'curr-program-loc'.

* In C++ translation semantics, rename 'curr-scope' to 'curr-tr-scope'.

* Rename some occurences of 'Name(arg1, arg2)' to 'ExecName(arg1, arg2)'.

In C++ translation semantics, 'Name' denotes a name which has NOT
been resolved yet, while in execution semantics, 'Name' denotes
already resolved block-local name (of a variable or a type).
But with 'constexpr', the translation semantics needs to distinguish between
the two meanings.

* In C++ translation semantics, rename 'env' cell to 'tr-env'.

* Rename 'types' to 'tr-types'.

* Rename 'block-control' to 'tr-block-control'.

* Rename 'block-stack' to 'tr-block-stack'

* Linking interface for 'curr-program-loc' and 'curr-scope'

* Linking interface for 'curr-tu'.

* C++ execution semantics: 'addToEnv' to 'addToExecEnv'

* rename CPP-ENV to CPP-EXECUTION-ENV

* CPP-EXECUTION-ENV

* WhileTStmt

* ForTStmt

* IfTStmt

* declareNonStaticObjectExec

* rename 'block-history' to 'tr-block-history'

* Rename 'this' cell to 'tr-this' in translation semantics.

* rename function 'currentTranslationUnit' to 'nameOfCurrentTranslationUnit'

* Refactor use of 'curr-tu' cell.

* Rename 'curr-tu' to 'curr-tr-tu' in C++ translation.

* Use 'thread-local' cell in C++ translation semantics.

* Move some parts of execution stack to 'thread-local' cell.

* compileTimeEval

* Translation(), Execution()

* generalize rewriting of [pre, xe, le] during execution

* TGotoStmt

* TpushBlock, TpopBlock

* TLabelStmt

* TCatchAnyOp

* TwriteVTables

* TaddToConstructed

* in C++ translation semantics, rename 'ReturnOp' to 'TReturnOp'

* TBreakStmt

* TExpressionStmt

* Mark more things as 'constexpr'

* Evaluate the initializer of local constexpr variable.

* 'initString' only in translation semantics

* Preserve <thread-local> cell while elaborating function declaration.

When elaborating function declaration, we create
'compile-time' local variables in <thread-local> cell
as shadows for all local constexpr variables.

In future we should do that for all const-qualified variables
whos initializer is a constant-expression.

* GitIgnore test results (*.kcc, *.kcc.out, and *.out files)

* Emit terms for execution (pushBlock etc)

* Extract common/execution for C and C++

* Rename C++ execution modules

* Rename C++ translation modules

* Create 'Execution()' / 'Translation()' guards

* Extract modules 'COMMON-IO-BUFFERED' and 'C-RACES'

* Extract module 'C-CPP-EXECUTION-EXPR-FUNCTION-CALL'.

Function call between C and C++ code is now in separate module.

This code is _not_ needed to support C++ compile-time function evaluation,
so it should not be included in C++ translation semantics.

* Extract module 'SIMPLE-EXECUTION' from C++ execution semantics

* Use Translation() / Execution() guards to prevent undesired rule application.

* Use some execution modules in C++ translation semantics to support constexpr.

* IO in C++ translation semantics to support compile-time function evaluation.

But without the OCaml bindings and side effects.

* Tests for constexpr.

* EPRVal - evaluated PRValue

This sort represents those PRValues that are fully evaluated
in the current Translation/Execution context. The terms of
pre(...) are considered fully evaluated only in Translation context,
but not in Execution context. In Execution context,
only prv(...) terms are considered fully evaluated.

The purpose of this is that we want to use the same 'context' declarations
for both Translation/Execution context. This way the contexts declared
in the language/execution directory do not apply in Translation context.

* Tests: add target for cleaning only C++ tests.

* ExprLoc

* KResult

* Constant expressions semantics for enumerators

* Destroy local constexpr variables created during the translation.

Otherwise they woud stay in <mem>, and since during the execution,
the integer part of keys are genarated sequentially from zero,
the obsolete keys would cause collisions.

* Do not use EPRVal without a real need.