Make size_t => unsigned-long
On my machine, there is a problem with the `new41.C` test. The test uses the `__SIZE_TYPE__` macro, that evaluates to `unsigned long`. Having `size_t => unsigned` in the semantics results in having two overloads for `operator new`, the implicit one with an `unsigned` parameter and the user-defined one with `unsigned long`, and the implicit one gets selected.